+++ src/job.rs
@@ -118,6 +118,8 @@ tagged_enum_or_default!(
             builds: Vec<ShortBuild>,
             /// HealthReport of the job
             health_report: Vec<HealthReport>,
+            /// SCM configured for the job
+            scm: SCM,
         },
         /// A pipeline project
         WorkflowJob (_class = "org.jenkinsci.plugins.workflow.job.WorkflowJob") {
@@ -168,6 +170,57 @@ tagged_enum_or_default!(
             /// Can resume blocked build
             resume_blocked: bool,
         },
+        /// A matrix project
+        MatrixProject (_class = "hudson.matrix.MatrixProject") {
+            /// Name of the job
+            name: String,
+            /// Display Name of the job
+            display_name: String,
+            /// Full Display Name of the job
+            full_display_name: String,
+            /// Full Name of the job
+            full_name: String,
+            /// Description of the job
+            description: String,
+            /// URL for the job
+            url: String,
+            /// Ball Color for the status of the job
+            color: BallColor,
+            /// Is the job buildable?
+            buildable: bool,
+            /// Is concurrent build enabled for the job?
+            concurrent_build: bool,
+            /// Are dependencies kept for this job?
+            keep_dependencies: bool,
+            /// Next build number
+            next_build_number: u32,
+            /// Is this job currently in build queue
+            in_queue: bool,
+            /// Link to the last build
+            last_build: Option<ShortBuild>,
+            /// Link to the first build
+            first_build: Option<ShortBuild>,
+            /// Link to the last stable build
+            last_stable_build: Option<ShortBuild>,
+            /// Link to the last unstable build
+            last_unstable_build: Option<ShortBuild>,
+            /// Link to the last successful build
+            last_successful_build: Option<ShortBuild>,
+            /// Link to the last unsucressful build
+            last_unsuccessful_build: Option<ShortBuild>,
+            /// Link to the last complete build
+            last_completed_build: Option<ShortBuild>,
+            /// Link to the last failed build
+            last_failed_build: Option<ShortBuild>,
+            /// List of builds of the job
+            builds: Vec<ShortBuild>,
+            /// HealthReport of the job
+            health_report: Vec<HealthReport>,
+            /// SCM configured for the job
+            scm: SCM,
+            /// Configurations for the job
+            active_configurations: Vec<ShortConfiguration>,
+        },
     }
 );
 
@@ -177,6 +230,7 @@ macro_rules! job_common_fields_dispatch {
             match self {
                 &Job::FreeStyleProject { ref $field, .. } => Ok($field),
                 &Job::WorkflowJob { ref $field, .. } => Ok($field),
+                &Job::MatrixProject { ref $field, .. } => Ok($field),
                 x @ &Job::Unknown { .. } => Err(client::Error::InvalidObjectType {
                     object_type: client::error::ExpectedType::Job,
                     action: client::error::Action::GetField(stringify!($field)),
@@ -191,6 +245,7 @@ macro_rules! job_common_fields_dispatch {
             match self {
                 &Job::FreeStyleProject { $field, .. } => Ok($field),
                 &Job::WorkflowJob { $field, .. } => Ok($field),
+                &Job::MatrixProject { $field, .. } => Ok($field),
                 x @ &Job::Unknown { .. } => Err(client::Error::InvalidObjectType {
                     object_type: client::error::ExpectedType::Job,
                     action: client::error::Action::GetField(stringify!($field)),
@@ -205,6 +260,7 @@ macro_rules! job_common_fields_dispatch {
             match self {
                 &Job::FreeStyleProject { ref $field, .. } => Ok($field),
                 &Job::WorkflowJob { ref $field, .. } => Ok($field),
+                &Job::MatrixProject { ref $field, .. } => Ok($field),
                 x @ &Job::Unknown { .. } => Err(client::Error::InvalidObjectType {
                     object_type: client::error::ExpectedType::Job,
                     action: client::error::Action::GetField(stringify!($field)),
@@ -367,4 +423,26 @@ pub struct HealthReport {
     icon_url: String,
     /// Score of the `Job`
     score: u16,
+}
+
+tagged_enum_or_default!(
+    /// An SCM
+    pub enum SCM {
+        /// No SCM
+        NullSCM (_class = "hudson.scm.NullSCM") {},
+        /// Git SCM
+        GitSCM (_class = "hudson.plugins.git.GitSCM") {},
+    }
+);
+
+/// A configuration of a matrix project
+#[derive(Debug, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct ShortConfiguration {
+    /// Name of the configuration, listing the values of the matrix
+    name: String,
+    /// Url to this configuration
+    url: String,
+    /// Ball Color for the status of the configuration
+    color: BallColor,
 }